backend/requirements.txt
fastapi[standard]
sqlalchemy
aiomysql
databases
cryptography
python-dotenv
pydantic-settings